fix: ae miss connection id
